Add plugin kernel tests
