+++ test/integration/plugin/CMakeLists.txt
@@ -11,13 +11,24 @@
 # Note: PLUGIN_BACKENDS is defined in ../CMakeLists.txt
 #
 foreach( BACKEND ${PLUGIN_BACKENDS} )
-  configure_file( test-plugin.cpp.in
-                  test-plugin-${BACKEND}.cpp )
-  raja_add_test( NAME test-plugin-${BACKEND}
-                 SOURCES ${CMAKE_CURRENT_BINARY_DIR}/test-plugin-${BACKEND}.cpp
+  configure_file( test-plugin-forall.cpp.in
+                  test-plugin-forall-${BACKEND}.cpp )
+  raja_add_test( NAME test-plugin-forall-${BACKEND}
+                 SOURCES ${CMAKE_CURRENT_BINARY_DIR}/test-plugin-forall-${BACKEND}.cpp
                          plugin_to_test.cpp )
 
-  target_include_directories(test-plugin-${BACKEND}.exe
+  target_include_directories(test-plugin-forall-${BACKEND}.exe
+                               PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/tests)
+endforeach()
+
+foreach( BACKEND ${PLUGIN_BACKENDS} )
+  configure_file( test-plugin-kernel.cpp.in
+                  test-plugin-kernel-${BACKEND}.cpp )
+  raja_add_test( NAME test-plugin-kernel-${BACKEND}
+                 SOURCES ${CMAKE_CURRENT_BINARY_DIR}/test-plugin-kernel-${BACKEND}.cpp
+                         plugin_to_test.cpp )
+
+  target_include_directories(test-plugin-kernel-${BACKEND}.exe
                                PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/tests)
 endforeach()
