+++ test/integration/plugin/tests/test-plugin.hpp
@@ -159,324 +159,4 @@ private:
   }
 };
 
-
-// Check that the plugin is called with the right Platform.
-// Check that the plugin is called the correct number of times,
-// once before and after each kernel capture for the capture counter,
-// once before and after each kernel invocation for the launch counter.
-
-// test with basic forall
-template <typename ExecPolicy,
-          typename WORKINGRES,
-          RAJA::Platform PLATFORM>
-void PluginForallTestImpl()
-{
-  SetupPluginVars spv(WORKINGRES{});
-
-  CounterData* data = plugin_test_resource->allocate<CounterData>(10);
-
-  for (int i = 0; i < 10; i++) {
-
-    RAJA::forall<ExecPolicy>(
-      RAJA::RangeSegment(i,i+1),
-      PluginTestCallable{data}
-    );
-
-    CounterData loop_data;
-    plugin_test_resource->memcpy(&loop_data, &data[i], sizeof(CounterData));
-    ASSERT_EQ(loop_data.capture_platform_active, PLATFORM);
-    ASSERT_EQ(loop_data.capture_counter_pre,     i+1);
-    ASSERT_EQ(loop_data.capture_counter_post,    i);
-    ASSERT_EQ(loop_data.launch_platform_active, PLATFORM);
-    ASSERT_EQ(loop_data.launch_counter_pre,     i+1);
-    ASSERT_EQ(loop_data.launch_counter_post,    i);
-  }
-
-  CounterData plugin_data;
-  plugin_test_resource->memcpy(&plugin_data, plugin_test_data, sizeof(CounterData));
-  ASSERT_EQ(plugin_data.capture_platform_active, RAJA::Platform::undefined);
-  ASSERT_EQ(plugin_data.capture_counter_pre,     10);
-  ASSERT_EQ(plugin_data.capture_counter_post,    10);
-  ASSERT_EQ(plugin_data.launch_platform_active, RAJA::Platform::undefined);
-  ASSERT_EQ(plugin_data.launch_counter_pre,     10);
-  ASSERT_EQ(plugin_data.launch_counter_post,    10);
-
-  plugin_test_resource->deallocate(data);
-}
-
-// test with basic forall_Icount
-template <typename ExecPolicy,
-          typename WORKINGRES,
-          RAJA::Platform PLATFORM>
-void PluginForAllICountTestImpl()
-{
-  SetupPluginVars spv(WORKINGRES{});
-
-  CounterData* data = plugin_test_resource->allocate<CounterData>(10);
-
-  for (int i = 0; i < 10; i++) {
-
-    RAJA::forall_Icount<ExecPolicy>(
-      RAJA::RangeSegment(i,i+1), i,
-      PluginTestCallable{data}
-    );
-
-    CounterData loop_data;
-    plugin_test_resource->memcpy(&loop_data, &data[i], sizeof(CounterData));
-    ASSERT_EQ(loop_data.capture_platform_active, PLATFORM);
-    ASSERT_EQ(loop_data.capture_counter_pre,     i+1);
-    ASSERT_EQ(loop_data.capture_counter_post,    i);
-    ASSERT_EQ(loop_data.launch_platform_active, PLATFORM);
-    ASSERT_EQ(loop_data.launch_counter_pre,     i+1);
-    ASSERT_EQ(loop_data.launch_counter_post,    i);
-  }
-
-  CounterData plugin_data;
-  plugin_test_resource->memcpy(&plugin_data, plugin_test_data, sizeof(CounterData));
-  ASSERT_EQ(plugin_data.capture_platform_active, RAJA::Platform::undefined);
-  ASSERT_EQ(plugin_data.capture_counter_pre,     10);
-  ASSERT_EQ(plugin_data.capture_counter_post,    10);
-  ASSERT_EQ(plugin_data.launch_platform_active, RAJA::Platform::undefined);
-  ASSERT_EQ(plugin_data.launch_counter_pre,     10);
-  ASSERT_EQ(plugin_data.launch_counter_post,    10);
-
-  plugin_test_resource->deallocate(data);
-}
-
-// test with IndexSet forall
-template <typename ExecPolicy,
-          typename WORKINGRES,
-          RAJA::Platform PLATFORM>
-void PluginForAllIdxSetTestImpl()
-{
-  SetupPluginVars spv(WORKINGRES{});
-
-  CounterData* data = plugin_test_resource->allocate<CounterData>(10);
-
-  for (int i = 0; i < 10; i++) {
-
-    RAJA::TypedIndexSet< RAJA::RangeSegment > iset;
-
-    for (int j = i; j < 10; j++) {
-      iset.push_back(RAJA::RangeSegment(j, j+1));
-    }
-
-    RAJA::forall<RAJA::ExecPolicy<RAJA::seq_segit, ExecPolicy>>(
-      iset,
-      PluginTestCallable{data}
-    );
-
-    for (int j = i; j < 10; j++) {
-      CounterData loop_data;
-      plugin_test_resource->memcpy(&loop_data, &data[j], sizeof(CounterData));
-      ASSERT_EQ(loop_data.capture_platform_active, PLATFORM);
-      ASSERT_EQ(loop_data.capture_counter_pre,     i+1);
-      ASSERT_EQ(loop_data.capture_counter_post,    i);
-      ASSERT_EQ(loop_data.launch_platform_active, PLATFORM);
-      ASSERT_EQ(loop_data.launch_counter_pre,     i+1);
-      ASSERT_EQ(loop_data.launch_counter_post,    i);
-    }
-  }
-
-  CounterData plugin_data;
-  plugin_test_resource->memcpy(&plugin_data, plugin_test_data, sizeof(CounterData));
-  ASSERT_EQ(plugin_data.capture_platform_active, RAJA::Platform::undefined);
-  ASSERT_EQ(plugin_data.capture_counter_pre,     10);
-  ASSERT_EQ(plugin_data.capture_counter_post,    10);
-  ASSERT_EQ(plugin_data.launch_platform_active, RAJA::Platform::undefined);
-  ASSERT_EQ(plugin_data.launch_counter_pre,     10);
-  ASSERT_EQ(plugin_data.launch_counter_post,    10);
-
-  plugin_test_resource->deallocate(data);
-}
-
-// test with IndexSet forall_Icount
-template <typename ExecPolicy,
-          typename WORKINGRES,
-          RAJA::Platform PLATFORM>
-void PluginForAllIcountIdxSetTestImpl()
-{
-  SetupPluginVars spv(WORKINGRES{});
-
-  CounterData* data = plugin_test_resource->allocate<CounterData>(10);
-
-  for (int i = 0; i < 10; i++) {
-
-    RAJA::TypedIndexSet< RAJA::RangeSegment > iset;
-
-    for (int j = i; j < 10; j++) {
-      iset.push_back(RAJA::RangeSegment(j, j+1));
-    }
-
-    RAJA::forall_Icount<RAJA::ExecPolicy<RAJA::seq_segit, ExecPolicy>>(
-      iset,
-      PluginTestCallable{data}
-    );
-
-    for (int j = i; j < 10; j++) {
-      CounterData loop_data;
-      plugin_test_resource->memcpy(&loop_data, &data[j], sizeof(CounterData));
-      ASSERT_EQ(loop_data.capture_platform_active, PLATFORM);
-      ASSERT_EQ(loop_data.capture_counter_pre,     i+1);
-      ASSERT_EQ(loop_data.capture_counter_post,    i);
-      ASSERT_EQ(loop_data.launch_platform_active, PLATFORM);
-      ASSERT_EQ(loop_data.launch_counter_pre,     i+1);
-      ASSERT_EQ(loop_data.launch_counter_post,    i);
-    }
-  }
-
-  CounterData plugin_data;
-  plugin_test_resource->memcpy(&plugin_data, plugin_test_data, sizeof(CounterData));
-  ASSERT_EQ(plugin_data.capture_platform_active, RAJA::Platform::undefined);
-  ASSERT_EQ(plugin_data.capture_counter_pre,     10);
-  ASSERT_EQ(plugin_data.capture_counter_post,    10);
-  ASSERT_EQ(plugin_data.launch_platform_active, RAJA::Platform::undefined);
-  ASSERT_EQ(plugin_data.launch_counter_pre,     10);
-  ASSERT_EQ(plugin_data.launch_counter_post,    10);
-
-  plugin_test_resource->deallocate(data);
-}
-
-// test with multi_policy forall
-template <typename ExecPolicy,
-          typename WORKINGRES,
-          RAJA::Platform PLATFORM>
-void PluginForAllMultiPolicyTestImpl()
-{
-  auto mp = RAJA::make_multi_policy<RAJA::seq_exec, ExecPolicy>(
-      [](const RAJA::RangeSegment &r) {
-        if (*(r.begin()) < 5) {
-          return 0;
-        } else {
-          return 1;
-        }
-      });
-
-  {
-    SetupPluginVars spv(camp::resources::Host{});
-
-    CounterData* data = plugin_test_resource->allocate<CounterData>(10);
-
-    for (int i = 0; i < 5; i++) {
-
-      RAJA::forall(mp,
-        RAJA::RangeSegment(i,i+1),
-        PluginTestCallable{data}
-      );
-
-      CounterData loop_data;
-      plugin_test_resource->memcpy(&loop_data, &data[i], sizeof(CounterData));
-      ASSERT_EQ(loop_data.capture_platform_active, RAJA::Platform::host);
-      ASSERT_EQ(loop_data.capture_counter_pre,     i+1);
-      ASSERT_EQ(loop_data.capture_counter_post,    i);
-      ASSERT_EQ(loop_data.launch_platform_active, RAJA::Platform::host);
-      ASSERT_EQ(loop_data.launch_counter_pre,     i+1);
-      ASSERT_EQ(loop_data.launch_counter_post,    i);
-    }
-
-    CounterData plugin_data;
-    plugin_test_resource->memcpy(&plugin_data, plugin_test_data, sizeof(CounterData));
-    ASSERT_EQ(plugin_data.capture_platform_active, RAJA::Platform::undefined);
-    ASSERT_EQ(plugin_data.capture_counter_pre,     5);
-    ASSERT_EQ(plugin_data.capture_counter_post,    5);
-    ASSERT_EQ(plugin_data.launch_platform_active, RAJA::Platform::undefined);
-    ASSERT_EQ(plugin_data.launch_counter_pre,     5);
-    ASSERT_EQ(plugin_data.launch_counter_post,    5);
-
-    plugin_test_resource->deallocate(data);
-  }
-
-  {
-    SetupPluginVars spv(WORKINGRES{});
-
-    CounterData* data = plugin_test_resource->allocate<CounterData>(10);
-
-    for (int i = 0; i < 5; i++) {
-
-      RAJA::forall(mp,
-        RAJA::RangeSegment(i+5,i+6),
-        PluginTestCallable{data}
-      );
-
-      CounterData loop_data;
-      plugin_test_resource->memcpy(&loop_data, &data[i+5], sizeof(CounterData));
-      ASSERT_EQ(loop_data.capture_platform_active, PLATFORM);
-      ASSERT_EQ(loop_data.capture_counter_pre,     i+1);
-      ASSERT_EQ(loop_data.capture_counter_post,    i);
-      ASSERT_EQ(loop_data.launch_platform_active, PLATFORM);
-      ASSERT_EQ(loop_data.launch_counter_pre,     i+1);
-      ASSERT_EQ(loop_data.launch_counter_post,    i);
-    }
-
-    CounterData plugin_data;
-    plugin_test_resource->memcpy(&plugin_data, plugin_test_data, sizeof(CounterData));
-    ASSERT_EQ(plugin_data.capture_platform_active, RAJA::Platform::undefined);
-    ASSERT_EQ(plugin_data.capture_counter_pre,     5);
-    ASSERT_EQ(plugin_data.capture_counter_post,    5);
-    ASSERT_EQ(plugin_data.launch_platform_active, RAJA::Platform::undefined);
-    ASSERT_EQ(plugin_data.launch_counter_pre,     5);
-    ASSERT_EQ(plugin_data.launch_counter_post,    5);
-
-    plugin_test_resource->deallocate(data);
-  }
-}
-
-TYPED_TEST_SUITE_P(PluginTest);
-template <typename T>
-class PluginTest : public ::testing::Test
-{
-};
-
-TYPED_TEST_P(PluginTest, PluginForall)
-{
-  using ExecPolicy = typename camp::at<TypeParam, camp::num<0>>::type;
-  using ResType = typename camp::at<TypeParam, camp::num<1>>::type;
-  using PlatformHolder = typename camp::at<TypeParam, camp::num<2>>::type;
-
-  PluginForallTestImpl<ExecPolicy, ResType, PlatformHolder::platform>( );
-}
-
-TYPED_TEST_P(PluginTest, PluginForAllICount)
-{
-  using ExecPolicy = typename camp::at<TypeParam, camp::num<0>>::type;
-  using ResType = typename camp::at<TypeParam, camp::num<1>>::type;
-  using PlatformHolder = typename camp::at<TypeParam, camp::num<2>>::type;
-
-  PluginForAllICountTestImpl<ExecPolicy, ResType, PlatformHolder::platform>( );
-}
-
-TYPED_TEST_P(PluginTest, PluginForAllIdxSet)
-{
-  using ExecPolicy = typename camp::at<TypeParam, camp::num<0>>::type;
-  using ResType = typename camp::at<TypeParam, camp::num<1>>::type;
-  using PlatformHolder = typename camp::at<TypeParam, camp::num<2>>::type;
-
-  PluginForAllIdxSetTestImpl<ExecPolicy, ResType, PlatformHolder::platform>( );
-}
-
-TYPED_TEST_P(PluginTest, PluginForAllIcountIdxSet)
-{
-  using ExecPolicy = typename camp::at<TypeParam, camp::num<0>>::type;
-  using ResType = typename camp::at<TypeParam, camp::num<1>>::type;
-  using PlatformHolder = typename camp::at<TypeParam, camp::num<2>>::type;
-
-  PluginForAllIcountIdxSetTestImpl<ExecPolicy, ResType, PlatformHolder::platform>( );
-}
-
-TYPED_TEST_P(PluginTest, PluginForAllMultiPolicy)
-{
-  using ExecPolicy = typename camp::at<TypeParam, camp::num<0>>::type;
-  using ResType = typename camp::at<TypeParam, camp::num<1>>::type;
-  using PlatformHolder = typename camp::at<TypeParam, camp::num<2>>::type;
-
-  PluginForAllMultiPolicyTestImpl<ExecPolicy, ResType, PlatformHolder::platform>( );
-}
-
-REGISTER_TYPED_TEST_SUITE_P(PluginTest,
-                            PluginForall,
-                            PluginForAllICount,
-                            PluginForAllIdxSet,
-                            PluginForAllIcountIdxSet,
-                            PluginForAllMultiPolicy);
-
 #endif  //__TEST_PLUGIN_HPP__
